backend/requirements.txt
# OmniData Backend Dependencies

# Core framework
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
python-multipart>=0.0.9

# LangGraph orchestration
langgraph>=0.2.0
langchain-core>=0.2.27,<0.3

# LLM inference
groq>=0.11.0

# Vector database
pinecone>=5.0.0

# Snowflake
snowflake-connector-python>=3.10.0

# SQL validation
sqlglot>=25.0.0

# E2B sandbox
e2b-code-interpreter==1.0.0

# Web search (Tavily)
tavily-python>=0.3.0

# Salesforce CRM
simple-salesforce>=1.12.0

# Data processing
pyyaml==6.0.2
pandas==2.2.0

# HTTP
httpx==0.27.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
